Refactor login and routes
