+++ asreview/webapp/src/Components/Header.js
@@ -1,55 +1,35 @@
-import React from "react";
-
+import { Menu, MenuOpen } from "@mui/icons-material";
+import { AppBar, Box, ButtonBase, IconButton, Toolbar } from "@mui/material";
 import { Link } from "react-router-dom";
-import { AppBar, Box, ButtonBase, Toolbar, IconButton } from "@mui/material";
-import { styled } from "@mui/material/styles";
-import { Menu } from "@mui/icons-material";
 
 import { ProfilePopper } from "Components";
 
 import { WordMark } from "icons/WordMark";
 
-const PREFIX = "Header";
-
-const classes = {
-  appBar: `${PREFIX}-appBar`,
-  menuButton: `${PREFIX}-menuButton`,
-  logo: `${PREFIX}-logo`,
-  toolbar: `${PREFIX}-toolbar`,
-};
-
-const Root = styled("div")(({ theme }) => ({
-  [`& .${classes.appBar}`]: {
-    [theme.breakpoints.up("md")]: {
-      zIndex: theme.zIndex.drawer + 1,
-    },
-  },
-  [`& .${classes.menuButton}`]: {
-    marginRight: 4,
-  },
-  [`& .${classes.logo}`]: {
-    width: 130,
-  },
-  [`& .${classes.toolbar}`]: {
-    justifyContent: "space-between",
-  },
-}));
-
-const Header = (props) => {
+const Header = ({ onNavDrawer, toggleNavDrawer, menuOpenButton = true }) => {
   return (
-    <Root aria-label="appbar-toolbar">
-      <AppBar color="inherit" position="fixed" className={classes.appBar}>
-        <Toolbar className={classes.toolbar}>
+    <>
+      <AppBar
+        color="inherit"
+        position="fixed"
+        sx={(theme) => ({
+          [theme.breakpoints.up("md")]: {
+            zIndex: theme.zIndex.drawer + 1,
+          },
+        })}
+      >
+        <Toolbar sx={{ justifyContent: "space-between" }}>
           <Box>
-            <IconButton
-              className={classes.menuButton}
-              edge="start"
-              color="inherit"
-              onClick={props.toggleNavDrawer}
-              size="large"
-            >
-              <Menu />
-            </IconButton>
+            {onNavDrawer !== undefined && (
+              <IconButton
+                edge="start"
+                color="inherit"
+                onClick={toggleNavDrawer}
+                size="large"
+              >
+                {menuOpenButton && onNavDrawer ? <MenuOpen /> : <Menu />}
+              </IconButton>
+            )}
             <ButtonBase
               disableRipple
               sx={{ width: "100px" }}
@@ -59,13 +39,12 @@ const Header = (props) => {
               <WordMark />
             </ButtonBase>
           </Box>
-          {window.authentication === true && (
-            <ProfilePopper mobilescreen={props.mobilescreen} />
-          )}
+          {window.authentication === true && <ProfilePopper />}
         </Toolbar>
       </AppBar>
       <Toolbar aria-label="placeholder toolbar" />
-    </Root>
+      {/* <Outlet /> */}
+    </>
   );
 };
