+++ asreview/webapp/src/Components/SignUpForm.js
@@ -1,61 +1,26 @@
-import React from "react";
-import { useNavigate } from "react-router-dom";
-import { useMutation } from "react-query";
 import LoadingButton from "@mui/lab/LoadingButton";
 import {
-  Box,
   Button,
-  Card,
   CardContent,
+  CardActions,
   Checkbox,
-  Fade,
+  Divider,
+  FormHelperText as FHT,
   FormControl,
   FormControlLabel,
-  FormHelperText as FHT,
   Stack,
   TextField,
   Typography,
-  Divider,
 } from "@mui/material";
+import { useMutation } from "react-query";
+import { useNavigate } from "react-router-dom";
 
-import { InlineErrorHandler } from ".";
-import { passwordRequirements, passwordValidation } from "globals.js";
-import { styled } from "@mui/material/styles";
-import { HelpPrivacyTermsButton } from "Components";
-import { useToggle } from "hooks/useToggle";
 import BaseAPI from "api/AuthAPI";
 import { useFormik } from "formik";
+import { passwordRequirements, passwordValidation } from "globals.js";
+import { useToggle } from "hooks/useToggle";
 import * as Yup from "yup";
-import { WordMark } from "icons/WordMark";
-
-const PREFIX = "SignUpForm";
-
-const classes = {
-  button: `${PREFIX}-button`,
-  card: `${PREFIX}-card`,
-  cardContent: `${PREFIX}-card-content`,
-};
-
-const Root = styled("div")(({ theme }) => ({
-  display: "flex",
-  height: "100%",
-  width: "100%",
-  alignItems: "center",
-  justifyContent: "center",
-  position: "absolute",
-  [`& .${classes.button}`]: {
-    paddingTop: theme.spacing(3),
-    paddingBottom: theme.spacing(3),
-    justifyContent: "space-between",
-  },
-  [`& .${classes.card}`]: {
-    borderRadius: theme.spacing(2),
-    width: "500px",
-  },
-  [`& .${classes.cardContent}`]: {
-    padding: "48px 40px !important",
-  },
-}));
+import { InlineErrorHandler } from ".";
 
 // VALIDATION SCHEMA
 const SignupSchema = Yup.object().shape({
@@ -70,7 +35,7 @@ const SignupSchema = Yup.object().shape({
     .oneOf([Yup.ref("password"), null], "Passwords must match"),
 });
 
-const SignUpForm = (props) => {
+const SignUpForm = ({ toggleSignUp }) => {
   // Pass the useFormik() hook initial form values, a validate function that will be called when
   // form values change or fields are blurred, and a submit function that will
   // be called when the form is submitted
@@ -96,11 +61,9 @@ const SignUpForm = (props) => {
     onSuccess: () => {
       let email = formik.values.email;
       formik.setValues(initialValues, false);
-      if (typeof props.showNotification === "function") {
-        props.showNotification(
-          `A confirmation email has been sent to ${email}.`,
-        );
-      }
+      // if (typeof showNotification === "function") {
+      //   showNotification(`A confirmation email has been sent to ${email}.`);
+      // }
       navigate("/signin");
     },
   });
@@ -111,10 +74,6 @@ const SignUpForm = (props) => {
     }
   };
 
-  const handleSignIn = () => {
-    navigate("/signin");
-  };
-
   const handleEnterKey = (e) => {
     if (e.keyCode === 13) {
       handleSubmit(e);
@@ -122,174 +81,162 @@ const SignUpForm = (props) => {
   };
 
   return (
-    <Root>
-      <Fade in>
-        <Box>
-          <Card className={classes.card} variant="outlined">
-            <CardContent className={classes.cardContent}>
-              <Stack spacing={3}>
-                <WordMark style={{ width: "100%", maxWidth: "130px" }} />
-                <Typography variant="h5">Create your profile</Typography>
-                <Stack spacing={3} component="form" noValidate>
-                  <TextField
-                    required={true}
-                    id="email"
-                    name="email"
-                    label="Email"
-                    size="small"
-                    type="email"
-                    autoComplete="email"
-                    fullWidth
-                    value={formik.values.email}
-                    onChange={formik.handleChange}
-                    onBlur={formik.handleBlur}
-                  />
-                  {formik.touched.email && formik.errors.email ? (
-                    <FHT error={true}>{formik.errors.email}</FHT>
-                  ) : null}
-                  <FormControl>
-                    <Stack direction="row" spacing={2}>
-                      <TextField
-                        required={true}
-                        id="password"
-                        label="Password"
-                        size="small"
-                        fullWidth
-                        type={showPassword ? "text" : "password"}
-                        value={formik.values.password}
-                        onChange={formik.handleChange}
-                        onBlur={formik.handleBlur}
-                        slotProps={{
-                          htmlInput: {
-                            autoComplete: "new-password",
-                          },
-                        }}
-                      />
-                      <TextField
-                        required={true}
-                        id="confirmPassword"
-                        label="Confirm Password"
-                        size="small"
-                        fullWidth
-                        type={showPassword ? "text" : "password"}
-                        onKeyDown={handleEnterKey}
-                        value={formik.values.confirmPassword}
-                        onChange={formik.handleChange}
-                        onBlur={formik.handleBlur}
-                        slotProps={{
-                          htmlInput: {
-                            autoComplete: "new-password",
-                          },
-                        }}
-                      />
-                    </Stack>
-                  </FormControl>
-
-                  <Typography
-                    variant="body2"
-                    sx={{ marginTop: "7px !important" }}
-                  >
-                    {passwordRequirements}
-                  </Typography>
-
-                  {formik.touched.password && formik.errors.password ? (
-                    <FHT error={true}>{formik.errors.password}</FHT>
-                  ) : null}
-                  {formik.touched.confirmPassword &&
-                  formik.errors.confirmPassword ? (
-                    <FHT error={true}>{formik.errors.confirmPassword}</FHT>
-                  ) : null}
-                  <FormControl>
-                    <FormControlLabel
-                      control={
-                        <Checkbox
-                          id="public"
-                          color="primary"
-                          onChange={toggleShowPassword}
-                        />
-                      }
-                      label="Show password"
-                    />
-                    {false && (
-                      <>
-                        <FormControlLabel
-                          control={
-                            <Checkbox
-                              color="primary"
-                              id="publicAccount"
-                              defaultChecked={formik.values.publicAccount}
-                              value={formik.values.publicAccount}
-                              onChange={formik.handleChange}
-                              onBlur={formik.handleBlur}
-                            />
-                          }
-                          label="Make this account public"
-                        />
-                        <FHT>
-                          Making this account public allows you to collaborate.
-                        </FHT>
-                      </>
-                    )}
-                  </FormControl>
-                  {isError && <InlineErrorHandler message={error.message} />}
-                  <Divider />
-                  <TextField
-                    required={true}
-                    id="name"
-                    name="name"
-                    label="Full name"
-                    size="small"
-                    autoComplete="name"
-                    fullWidth
-                    value={formik.values.name}
-                    onChange={formik.handleChange}
-                    onBlur={formik.handleBlur}
-                  />
-                  {formik.touched.name && formik.errors.name ? (
-                    <FHT error={true}>{formik.errors.name}</FHT>
-                  ) : null}
-                  <TextField
-                    required={true}
-                    id="affiliation"
-                    label="Affiliation"
-                    size="small"
-                    name="affiliation"
-                    autoComplete="organization"
-                    fullWidth
-                    value={formik.values.affiliation}
-                    onChange={formik.handleChange}
-                    onBlur={formik.handleBlur}
-                  />
-                  {formik.touched.affiliation && formik.errors.affiliation ? (
-                    <FHT error={true}>{formik.errors.affiliation}</FHT>
-                  ) : null}
-                  <Stack className={classes.button} direction="row">
-                    <Button
-                      id="sign-in"
-                      onClick={handleSignIn}
-                      sx={{ textTransform: "none" }}
-                    >
-                      Sign In instead
-                    </Button>
-                    <LoadingButton
-                      //loading={isLoading}
-                      id="create-profile"
-                      variant="contained"
-                      color="primary"
-                      onClick={handleSubmit}
-                      disabled={!(formik.isValid && formik.dirty)}
-                    >
-                      Create
-                    </LoadingButton>
-                  </Stack>
-                </Stack>
+    <>
+      <CardContent>
+        <Stack spacing={3}>
+          <Typography variant="h5">Create your profile</Typography>
+          <Stack spacing={3} component="form" noValidate>
+            <TextField
+              required={true}
+              id="email"
+              name="email"
+              label="Email"
+              size="small"
+              type="email"
+              autoComplete="email"
+              fullWidth
+              value={formik.values.email}
+              onChange={formik.handleChange}
+              onBlur={formik.handleBlur}
+            />
+            {formik.touched.email && formik.errors.email ? (
+              <FHT error={true}>{formik.errors.email}</FHT>
+            ) : null}
+            <FormControl>
+              <Stack direction="row" spacing={2}>
+                <TextField
+                  required={true}
+                  id="password"
+                  label="Password"
+                  size="small"
+                  fullWidth
+                  type={showPassword ? "text" : "password"}
+                  value={formik.values.password}
+                  onChange={formik.handleChange}
+                  onBlur={formik.handleBlur}
+                  slotProps={{
+                    htmlInput: {
+                      autoComplete: "new-password",
+                    },
+                  }}
+                />
+                <TextField
+                  required={true}
+                  id="confirmPassword"
+                  label="Confirm Password"
+                  size="small"
+                  fullWidth
+                  type={showPassword ? "text" : "password"}
+                  onKeyDown={handleEnterKey}
+                  value={formik.values.confirmPassword}
+                  onChange={formik.handleChange}
+                  onBlur={formik.handleBlur}
+                  slotProps={{
+                    htmlInput: {
+                      autoComplete: "new-password",
+                    },
+                  }}
+                />
               </Stack>
-            </CardContent>
-          </Card>
-          <HelpPrivacyTermsButton />
-        </Box>
-      </Fade>
-    </Root>
+            </FormControl>
+
+            <Typography variant="body2" sx={{ marginTop: "7px !important" }}>
+              {passwordRequirements}
+            </Typography>
+
+            {formik.touched.password && formik.errors.password ? (
+              <FHT error={true}>{formik.errors.password}</FHT>
+            ) : null}
+            {formik.touched.confirmPassword && formik.errors.confirmPassword ? (
+              <FHT error={true}>{formik.errors.confirmPassword}</FHT>
+            ) : null}
+            <FormControl>
+              <FormControlLabel
+                control={
+                  <Checkbox
+                    id="public"
+                    color="primary"
+                    onChange={toggleShowPassword}
+                  />
+                }
+                label="Show password"
+              />
+              {false && (
+                <>
+                  <FormControlLabel
+                    control={
+                      <Checkbox
+                        color="primary"
+                        id="publicAccount"
+                        defaultChecked={formik.values.publicAccount}
+                        value={formik.values.publicAccount}
+                        onChange={formik.handleChange}
+                        onBlur={formik.handleBlur}
+                      />
+                    }
+                    label="Make this account public"
+                  />
+                  <FHT>
+                    Making this account public allows you to collaborate.
+                  </FHT>
+                </>
+              )}
+            </FormControl>
+            {isError && <InlineErrorHandler message={error.message} />}
+            <Divider />
+            <TextField
+              required={true}
+              id="name"
+              name="name"
+              label="Full name"
+              size="small"
+              autoComplete="name"
+              fullWidth
+              value={formik.values.name}
+              onChange={formik.handleChange}
+              onBlur={formik.handleBlur}
+            />
+            {formik.touched.name && formik.errors.name ? (
+              <FHT error={true}>{formik.errors.name}</FHT>
+            ) : null}
+            <TextField
+              required={true}
+              id="affiliation"
+              label="Affiliation"
+              size="small"
+              name="affiliation"
+              autoComplete="organization"
+              fullWidth
+              value={formik.values.affiliation}
+              onChange={formik.handleChange}
+              onBlur={formik.handleBlur}
+            />
+            {formik.touched.affiliation && formik.errors.affiliation ? (
+              <FHT error={true}>{formik.errors.affiliation}</FHT>
+            ) : null}
+          </Stack>
+        </Stack>
+      </CardContent>
+      <CardActions>
+        <Button
+          id="sign-in"
+          onClick={toggleSignUp}
+          sx={{ textTransform: "none" }}
+        >
+          Sign In instead
+        </Button>
+        <LoadingButton
+          //loading={isLoading}
+          id="create-profile"
+          variant="contained"
+          color="primary"
+          onClick={handleSubmit}
+          disabled={!(formik.isValid && formik.dirty)}
+        >
+          Create
+        </LoadingButton>
+      </CardActions>
+    </>
   );
 };
