+++ asreview/webapp/src/Components/index.js
@@ -1,24 +1,22 @@
+export { default as AuthPage } from "./AuthPage";
 export { default as BoxErrorHandler } from "./BoxErrorHandler";
 export { default as CardErrorHandler } from "./CardErrorHandler";
 export { default as CiteDialog } from "./CiteDialog";
-export { default as ConfirmAccount } from "./ConfirmAccount";
 export { default as DialogErrorHandler } from "./DialogErrorHandler";
-export { default as DrawerItemContainer } from "./DrawerItemContainer";
 export { default as ElasGame } from "./ElasGame";
 export { default as ForgotPassword } from "./ForgotPassword";
 export { default as Header } from "./Header";
 export { default as HelpDialog } from "./HelpDialog";
 export { default as HelpPrivacyTermsButton } from "./HelpPrivacyTermsButton";
 export { default as InlineErrorHandler } from "./InlineErrorHandler";
-export { default as NavigationDrawer } from "./NavigationDrawer";
+export { default as InteractionButtons } from "./InteractionButtons";
+export { default as LandingDrawerItems } from "./LandingDrawerItems";
 export { default as OpenInNewIconStyled } from "./OpenInNewIconStyled";
-export { default as PersistSignIn } from "./PersistSignIn";
+export { default as PageWithDrawer } from "./PageWithDrawer";
 export { default as ProfilePopper } from "./ProfilePopper";
+export { default as ProjectDrawerItems } from "./ProjectDrawerItems";
 export { default as ResetPassword } from "./ResetPassword";
-export { default as RequireAuth } from "./RequireAuth";
 export { default as SettingsDialog } from "./SettingsDialog";
-export { default as SignIn } from "./SignIn";
 export { default as SignInForm } from "./SignInForm";
 export { default as SignInOAuthCallback } from "./SignInOAuthCallback";
 export { default as SignUpForm } from "./SignUpForm";
-export { default as InteractionButtons } from "./InteractionButtons";
