--- asreview/webapp/src/HomeComponents/index.js
@@ -1,1 +1,0 @@
-export { default as HomePage } from "./HomePage";
