+++ asreview/webapp/src/ProjectComponents/ImportProject.js
@@ -15,7 +15,7 @@ import {
   Stack,
   Typography,
 } from "@mui/material";
-import { styled, useTheme } from "@mui/material/styles";
+import { styled } from "@mui/material/styles";
 import React, { useCallback, useMemo } from "react";
 import { useDropzone } from "react-dropzone";
 import { useMutation, useQueryClient } from "react-query";
@@ -140,8 +140,7 @@ const ImportProjectCard = ({ onClose = null, mutate, isLoading }) => {
 const ImportProject = ({ ...buttonProps }) => {
   const queryClient = useQueryClient();
 
-  const theme = useTheme();
-  const smallScreen = useMediaQuery(theme.breakpoints.down("sm"));
+  const smallScreen = useMediaQuery((theme) => theme.breakpoints.down("sm"));
 
   const [importSnackbar, toggleImportSnackbar] = useToggle();
   const [warningDialog, toggleWarningDialog] = useToggle();
