+++ asreview/webapp/src/ProjectComponents/index.js
@@ -1,3 +1,2 @@
 export { default as ImportProject } from "./ImportProject";
 export { default as ProjectDeleteDialog } from "./ProjectDeleteDialog";
-export { default as ProjectPage } from "./ProjectPage";
