+++ asreview/webapp/src/RouteNotFound.js
@@ -1,42 +1,18 @@
-import * as React from "react";
+import { Box, Button, Fade, Stack, Typography } from "@mui/material";
 import { useNavigate } from "react-router-dom";
-import { Button, Fade, Stack, Typography } from "@mui/material";
-import { styled } from "@mui/material/styles";
-
 import ElasPotter from "./images/ElasPotter.svg";
 
-const PREFIX = "RouteNotFound";
-
-const classes = {
-  img: `${PREFIX}-img`,
-  textTitle: `${PREFIX}-textTitle`,
-};
-
-const Root = styled("div")(({ theme }) => ({
-  height: "inherit",
-  [`& .${classes.img}`]: {
-    maxWidth: 250,
-    [theme.breakpoints.down("md")]: {
-      maxWidth: 150,
-    },
-  },
-  [`& .${classes.textTitle}`]: {
-    textAlign: "center",
-    [theme.breakpoints.down("md")]: {
-      width: "80%",
-    },
-  },
-}));
-
-const RouteNotFound = (props) => {
+const RouteNotFound = () => {
   const navigate = useNavigate();
 
-  const handleClickHomePage = () => {
-    navigate("/");
-  };
-
   return (
-    <Root aria-label="route not found">
+    <Box
+      aria-label="route not found"
+      display="flex"
+      justifyContent="center"
+      alignItems="center"
+      minHeight="100vh"
+    >
       <Fade in>
         <Stack
           spacing={3}
@@ -46,16 +22,16 @@ const RouteNotFound = (props) => {
             justifyContent: "center",
           }}
         >
-          <img src={ElasPotter} alt="ElasPotter" className={classes.img} />
+          <img src={ElasPotter} alt="ElasPotter" style={{ width: "240px" }} />
           <Stack spacing={3} sx={{ alignItems: "center" }}>
-            <Typography className={classes.textTitle} variant="h5">
-              Whoops! You have come to an unexplored land.
+            <Typography variant="h5">
+              Whoops! You have come to an unexplored land named 404
             </Typography>
-            <Button onClick={handleClickHomePage}>Go home</Button>
+            <Button onClick={() => navigate("/reviews")}>Take me back</Button>
           </Stack>
         </Stack>
       </Fade>
-    </Root>
+    </Box>
   );
 };
