+++ asreview/webapp/src/StyledComponents/StyledList.js
@@ -1,0 +1,9 @@
+import { List } from "@mui/material";
+import { styled } from "@mui/material/styles";
+
+export const StyledList = styled(List)(({ theme }) => ({
+  overflow: "hidden",
+  display: "flex",
+  flexDirection: "column",
+  flexGrow: 1,
+}));
