+++ asreview/webapp/tests/utils/api_utils.py
@@ -66,13 +66,6 @@ def signout_user(client):
     return response
 
 
-def confirm_user(client, user):
-    response = client.post(
-        "/auth/confirm_account", data={"user_id": user.id, "token": user.token}
-    )
-    return response
-
-
 def forgot_password(client, user):
     response = client.post("/auth/forgot_password", data={"email": user.email})
     return response
